back/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.2.1
Faker==37.6.0
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.60.0
fqdn==1.5.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.5
idna==3.10
ipykernel==6.30.1
ipython==8.17.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
prometheus_client==0.22.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyMuPDF==1.22.5
pyparsing==3.2.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.6
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.201
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.23.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
easyocr==1.7.2
paddleocr==3.2.0
